rename `clean_caches` to `clear_taichi_aot_caches`, `check_kernels_count` to `count_taichi_aot_kernels`

+++ brainpy/_src/math/op_register/taichi_aot_based.py
@@ -8,7 +8,7 @@ import platform
 import re
 import shutil
 from functools import partial, reduce
-from typing import Any, Sequence
+from typing import Any, Sequence, Union
 
 import jax.core
 import numpy as np
@@ -16,14 +16,17 @@ from jax.interpreters import xla, mlir
 from jax.lib import xla_client
 from jaxlib.hlo_helpers import custom_call
 
-from brainpy.errors import PackageMissingError
 from brainpy._src.dependency_check import (import_taichi,
                                            import_brainpylib_cpu_ops,
                                            import_brainpylib_gpu_ops)
+from brainpy.errors import PackageMissingError
 from .utils import _shape_to_layout
 
 
-### UTILS ###
+taichi_cache_path = None
+
+
+# --- UTILS ###
 
 # get the path of home directory on Linux, Windows, Mac
 def get_home_dir():
@@ -43,8 +46,18 @@ def encode_md5(source: str) -> str:
 
   return md5.hexdigest()
 
+
 # check kernels count
-def check_kernels_count() -> int:
+def count_taichi_aot_kernels() -> int:
+  """
+  Count the number of AOT compiled kernels.
+
+  Returns
+  -------
+  kernels_count: int
+    The number of AOT compiled kernels.
+
+  """
   if not os.path.exists(kernels_aot_path):
     return 0
   kernels_count = 0
@@ -54,23 +67,37 @@ def check_kernels_count() -> int:
     kernels_count += len(dir2)
   return kernels_count
 
-# clean caches
-def clean_caches(kernels_name: list[str]=None):
-  if kernels_name is None:
-    if not os.path.exists(kernels_aot_path):
-      raise FileNotFoundError("The kernels cache folder does not exist. \
-                              Please define a kernel using `taichi.kernel` \
-                              and customize the operator using `bm.XLACustomOp` \
-                              before calling the operator.")
-    shutil.rmtree(kernels_aot_path)
-    print('Clean all kernel\'s cache successfully')
+
+def clear_taichi_aot_caches(kernels: Union[str, Sequence[str]] = None):
+  """
+  Clean the cache of the AOT compiled kernels.
+  
+  Parameters
+  ----------
+  kernels: str or list of str
+    The name of the kernel to be cleaned. If None, all the kernels will be cleaned.
+  """
+  if kernels is None:
+    global taichi_cache_path
+    if taichi_cache_path is None:
+      from taichi._lib.utils import import_ti_python_core
+      taichi_cache_path = import_ti_python_core().get_repo_dir()
+    # clean taichi cache
+    if os.path.exists(taichi_cache_path):
+      shutil.rmtree(taichi_cache_path)
+    # clean brainpy-taichi AOT cache
+    if os.path.exists(kernels_aot_path):
+      shutil.rmtree(kernels_aot_path)
     return
-  for kernel_name in kernels_name:
-    try:
+  if isinstance(kernels, str):
+    kernels = [kernels]
+  if not isinstance(kernels, list):
+    raise TypeError(f'kernels_name must be a list of str, but got {type(kernels)}')
+  # clear brainpy kernel cache
+  for kernel_name in kernels:
+    if os.path.exists(os.path.join(kernels_aot_path, kernel_name)):
       shutil.rmtree(os.path.join(kernels_aot_path, kernel_name))
-    except FileNotFoundError:
-      raise FileNotFoundError(f'Kernel {kernel_name} does not exist.')
-  print('Clean kernel\'s cache successfully')
+
 
 # TODO
 # not a very good way
@@ -104,7 +131,7 @@ def is_metal_supported():
   return True
 
 
-### VARIABLES ###
+# --- VARIABLES ###
 home_path = get_home_dir()
 kernels_aot_path = os.path.join(home_path, '.brainpy', 'kernels')
 is_metal_device = is_metal_supported()
@@ -122,7 +149,7 @@ def _check_kernel_exist(source_md5_encode: str) -> bool:
     return False
 
 
-### KERNEL AOT BUILD ###
+# --- KERNEL AOT BUILD ###
 
 
 def _array_to_field(dtype, shape) -> Any:
@@ -212,7 +239,7 @@ def _build_kernel(
   kernel.__name__ = kernel_name
 
 
-### KERNEL CALL PREPROCESS ###
+# --- KERNEL CALL PREPROCESS ###
 
 # convert type to number
 type_number_map = {
@@ -334,9 +361,6 @@ def _preprocess_kernel_call_gpu(
   return opaque
 
 
-
-
-
 def _XlaOp_to_ShapedArray(c, xla_op):
   xla_op = c.get_shape(xla_op)
   return jax.core.ShapedArray(xla_op.dimensions(), xla_op.element_type())
@@ -376,7 +400,7 @@ def _compile_kernel(abs_ins, kernel, platform: str, **kwargs):
       try:
         os.removedirs(os.path.join(kernels_aot_path, source_md5_encode))
       except Exception:
-          raise RuntimeError(f'Failed to preprocess info to build kernel:\n\n {codes}') from e
+        raise RuntimeError(f'Failed to preprocess info to build kernel:\n\n {codes}') from e
       raise RuntimeError(f'Failed to build kernel:\n\n {codes}') from e
 
   # returns
